% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_missclass.R
\name{correct_misclass}
\alias{correct_misclass}
\title{Function to correct misclassification based on \code{nonprobsvy} object}
\usage{
correct_misclass(
  obj,
  y,
  x,
  y_class_mat,
  x_class_mat,
  valid_sample,
  svy,
  method = c("ML", "simex", "imputation")
)
}
\arguments{
\item{obj}{\code{nonprobsvy} object (class: \code{nonprobsvy})}

\item{y}{either a formula with variables that are misclassified or a list of objects if validation sample is provided}

\item{x}{either a formula with variables that are misclassified or a list of objects if validation sample is provided}

\item{y_class_mat}{either a single matrix or a list of misclassification matrices for \code{y} variables}

\item{x_class_mat}{either a single matrix or a list of misclassification matrices for \code{x} variables}

\item{valid_sample}{a data.frame with validation sample where columns are determined by \code{y} and \code{x} arguments}

\item{svy}{a \code{survey.design2} object}

\item{method}{a vector of possible methods to correct for the misclassification}
}
\value{
An \code{nonprobsvy} object corrected for misclassification
}
\description{
Function to correct misclassification based on \code{nonprobsvy} object
}
\details{
Details will be provided here
}
\examples{
# examples will be here

}
