% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mc_variable_name}
\alias{.mc_variable_name}
\title{Name of the variable inside the mc() term of a formula}
\usage{
.mc_variable_name(formula)
}
\description{
Name of the variable inside the mc() term of a formula
}
\keyword{internal}
